test(mistral): merge fragmented-name tests back into mistral.test.ts
Splitting into a separate test file caused mock conflicts under
vitest's non-isolated mode (isolate: false), resulting in 401
Invalid API Key errors because vi.mock for @mistralai/mistralai
was not consistently intercepted across files.

Merge the 15 regression tests back into mistral.test.ts and add
the standard oxlint-disable max-lines grandfather comment used
elsewhere in the repo for oversized test files.

Also run oxfmt to fix formatting.
